Added Hash to plist file names
This is done in order to support files with the same in diffrent
directories. Added simple cli test for --plist-output cmd.
Added fix for #8289 an error is generated if the --plist-output
folder does not exist.
